Fix day 5 part 2 solution
The brute forcing solution to the day 5 part 2 challenge did not
complete in a reasonable timeframe.

Fix this by taking a different approach involving mapping the seed
ranges rather than individual seed numbers. This involves some
complicated logic to solve the problem that some of the seed ranges
overlap with some of the almanac ranges. The solution to that is to
split the seed ranges at the intersection point with the almanac range,
map the part of the seed range that coincides with the almanac range,
then add the part that doesn't to an `unmapped` array that is processed
later.

This now finds the correct answer in just 0.25 seconds.